fix(library.properties): Remove redundant `url` property from library.properties
Previously the library.properties metadata file contained two `url` property definitions. In this case, the later definition overrides the previous one.

The first definition correctly set the property to the repository URL. However, the second one overrode that correct value with an empty value, which caused a rejection of the submission of the library to the Arduino Library Manager registry:

> ERROR: library.properties url value is less than minimum length.